Complete `media` tests and docstrings Split `player` tests to separate test file
